Reorderchild and insert child optimizations
Now for every added element and reordered element a linear search is performed and items in the children and/or descendants array are shifted when the new/reorderd child is added.

This implementation batches all the changes and does one insertion sort on the children array, in case of CCSpriteBatchNode it also gives each child a new atlasIndex and performs one insertion sort on the descendants array.

What has changed?
- CCNode, addChild, reorderChild are changed
sortAllChildren is added
isReorderChildDirty_ is added
visit is changed

- CCSprite
sortAllChildren is overriden
reorderChild and addChild are changed
setReorderChildDirtyRecursively is added

- CCSpriteBatchNode
- addChild and reorderChild are changed
- appendChild, updateAtlasIndexes, sortDescendants are added
- visit is slightly changed, sortAllChildren is changed

What does it do?

- CCNode, a added child is appened to the children array, a reorderd child z-order is set immediately. Sorting happens before drawing. A dirty bool keeps track if sorting is necessary or not.

- CCSprite, overrides sortChildren, when it is added to a batchNode it also sorts its children recursively.

-CCSpriteBatchNode, a new child is appended to the descendants array as well. Sorting is 3 stages. First recursively all children are sorted based on their zOrder. Second every child gets a new atlasIndex, based on their relative zOrder (keeping parent -> child relations intact). Third all descendants and quads are sorted based on the atlasIndex.

All intense looping methods are optimized with fast dispatch.


Former-commit-id: 36f3dcf35c7534a47a55e0b5d22817702d8f87de