Implemented cascading opacity feature. Added CCLayerRGBA and CCLayerColor cascading opacity tests to LayerTest.
Adds CCNodeRGBA and CCLayerRGBA. Both conform to the CCRGBAProtocol and propagate opacity to their children. CCSprite now inherits from CCNodeRGBA, and CCMenu from CCLayerRRGBA.

Enable/disable this feature using CC_CASCADING_OPACITY in ccConfig.h.

When this feature is enabled, all classes calculate their cumulative opacity from their own opacity and from parent's opacity. For example, if you create a CCLayerRGB (or any other CCNodeRGB0 with an opacity of 127 (50%), and then add a children CCSprite with an opacity of 127, the sprite draws as if having an opacity of 63 (25%).


Former-commit-id: 6be962f8d08a4d76b486b1fa128ba51ec5b57a31